created the audio context with single tone
-AudioContext- is the object which we interact with the Web Audio API. Next we create our first
node, an oscillator. -- In electronics and signal processing terms, an _oscillator_ is a device that
creates a signal that repeatedly goes up and down in a regular pattern.
In this example, we're setting the frequency to 440 Hertz (Hz), or 440 cycles per second. In other words,
the oscillator is outputting a signal that transitions from 0 to 1 to -1 and back to 0 a total of 440 times
every second. This means that one cycle of the wave lasts 1/440 of a second, or 2.27ms.

440 Hz is the standard reference pitch for tuning musical instruments. The frequency corresponds to the
note A above middle C.

Next, we create a gain node and set its value to 0.5. In signal processing, gain refers to an increase or
decrease in a signal's amplitude, or its range of values. -- In practical terms, gain acts as a volume control.
A gain of 2 doubles the amplitude, making the sound louder, a gain of 0.5 halves the amplitude, making the sound
softer, and a gain of 1 (the default value of a gain node) has no effect on the amplitude.

So far we have two nodes: an oscillator node and a gain node. To actually apply the gain to the oscillators signal,
we need to connect the nodes together. We link the output of the oscillator node to the input of the gain node
using the oscillator's -connect- method.

Then, to be able to hear the result, we connect the output of the gain node to main output, which is available to
us through the audio context as -ctx.destination- . -- These connections mean that the oscillator signal is passed
through the gain node and then passed to the output, which will ultimately go to your headphones or speakers, if
the sound is turned on.

To illustrate these connections:
[Oscillator {frequency: 440}] output -> [Gain {gain=0.5}] output -> [Output] MAIN_OUTPUT -> to speakers or headphones

The oscillator node doesn't actually create a signal until we tell it to. - To do that, we call the -start- method on
the ocillator, passing -audioCtx.currentTime- as an agrument -- The -currentTime- property corresponds to the amount
of time in seconds that the audio context has been active.

By passing the -autdioCtx.currentTime- to the -start- method, we're telling the osciallator to start playing immediately.
Then we call the -stop- method, passing -audioCtx.currentTime + 2- -- This tells the ocillator to stop two seconds after
it started.