Created a call to tone.js with -Tone.start()-
-Tone.start- triggers the Tone.js library to start inside the click handler, ensuring that
the browser will allow it to play audio.

Next, a new -Tone.Synth- object is created. Synth is short for _synthesizer_ , an electronic
instrument, usually with a keyboard, that can generate(synthesize) all kinds of sounds. A
-Tone.Synth- is a simple code version of such an instrument.

The -Tone.Synth- constructor takes an object as its argument that allows us to configure various
aspects of the synthesizer. -- In this case, we're telling the synth to use oscillator that
generates sine waves. We're also giving the synth a simple amplitude envelope and a volume of -6.

After the constructor we chain the -toDestination- method, which connects the output of the synth
to the audio context's output

Finally, we tell the synth to play a single note using its -tiggerAttackRelease- method. -- This
method takes the note's name, the duration , and the time at which to play the note. --- We're passing
{A4} for the note name, which is equivalent to 440 Hz, and telling it to play for two seconds, starting
immediately.