updated the -envelope- property of -Tone.Synth- instance wrapped in -synth- variable
the second argument to -triggerAttackRelease- is the duration of the note in seconds, and the third
argument is the time at which to play the note, also in seconds. -- The first argument is the note
to be played. Starting from A3 -> B3 -> C#4 -> ... A4

Each note will play for 1 second total in sequence with no overlap between notes