replaced -Tone.Synth- object with -Tone.PolySynth-
The -Tone.PolySynth- constructor takes two arguments: a monophonic synth (in this case, -Tone.Synth-)
and an object with the options that would normally be passed to that monophonic synth's constructor.

The polysynth then screates multiple monophonic synths with the specified settings, effectively allowing
it to play multiple notes at once.

Next, we play the same scale but with addtional simultaneous notes. This is achieved by passing an array of
note names to the -tiggerAttackRelease- method instead of a single note name.
 For Example: We pass the array [A3, C#4] to play A3 and C#4 as the same time.

When you play this example, you should hear a harmonized scale with a nice magor chord at the end.