updated -synth- and created -Tone.Loop- instance
For the synth instance; we are not passing any arguments, meaning it will use the library's defaults (Tone.js = library)

Next, we create a new instance of -Tone.Loop- which has a constructor with two arguments. The first agruments is some
function that requires a time value, and the second argument is a duration indicating how often to call the function in
the first argument.

In this case, we pass the string [4n] as the second argument, which is Tone.js's notation for a quarter note
 ( [4n] is short for [1/4 note] ) -- This means the loop will repeat every beat.

The body of the callback function we pass to -Tone.Loop- calss the -triggerAttackRelease- method on the synth to play the note
C4 with a sixteenth note duration.

The third argument of the -triggerAttackRelase- method, -time- ; Represents the time to play a note. The -Tone.Loop- object will
provide a new -time- value whenever it calls the callback function, filling it in with the approriate location on the transport.

Finally, we call the start and stop methods on the loop returned by the -Tone.Loop- constructor; -- Passing the time when we want
this loop to start and when we want it to stop, followed by a call to -Tone.Transport.start- ; Which starts the transport playing
from the beginning.

Our loop repeats at every beat, playing one note each time, so we'll play a total of 16 notes.