For this example, we're switching to a polysynth
This is so we can play multiple notes at once. The _notes_ array contains one octave of a C major pentatonic scale,
including the C from the next octave.

Inside the -Tone.Loop- callback, we use a -for- loop to run some code three times. -- For each time around, we call
Math.random() ; which returns a random number between 0 and 1; to determine whether to play a note or not. If the value
is less than 0.5, we play a note. -- Otherwise, that note is skipped.

The _note name_ is determined by picking a random index into the _notes_ array, using the code -Mathfloor.(Math.random() * notes.length)-