updated -synth- instance
First, the callback function takes two agruments, _time_ and _note_ -- Instead of a single time argument.
Second, there's an extra argument after the callback, which contains a list of notes. Each time the callback
is called, the next note in this list is passed as the _note_ argument. It will keep cycling through the notes
in the list over and over until it's time to stop. -- The third argument to -Tone.Sequence- gives the duration
between each callback. - In this case we've used '4n', which means that a new note will be played every quarter
note.

When you run this example, you should hear a pattern play for 4 bars, with 4 beats per bar, making 16 notes in total.