changed the -synth- and created a -Tone.Part- object
The first change we're making here is to the synth. Using a polyphonic synth, so we can play multiple notes at
the same time. -- Other than the synth being different, the body of the callback function is the same. We're
still calling -synth.triggerAttackRelease- and passing the _note_ and _time_ parameters, -- Which -Tone.Part-
will fill in automatically.

Next comes the array of time/note pairs. Notice that some of the notes are arrays themselves; This pair of notes
will be passed to the -triggerAttackRelease- method unchanged and will have the effect of play two notes at once,

Finally, we call -.start('0:0:0')- which has the effect of playing this part immediately. If we used -.start('1:0:0')
instead, for example, then the melody would start after a bar's pause.

The times given for each time/note pair are relative to the time passed to the start method.

When you play this example, you should hear the beginning of 'Mary Had a Little Lamb'