updated the README.md -- created new -Tone.NoiseSynth-
Created a new -Tone.NoiseSynth- wrapped in -hiHat- passing an amplitude envelope and a volume of -6 dB. The envelope
has a very short attack (1/1000 of a second) and a longer decay (1/10 of a second), which is supposed to mimic the amplitude
envelope of a hi-hat being struck. Because _sustain_ and _relase_ are both set to 0, the sound will be over immediately after
the initial attack and decay periods (0.001 + 0.1 s).

In particular, the _sustain_ of 0 means the sound will sustain at 0 percent of its full volume, so even if the duration of the
note is longer, you won't hear anything after the attack and decay.

Next, we use -Tone.Loop- to play a continuous stream of eighth-note hi-hats for four bars. -- Note that the -triggerAttackRelease- method
on -NoiseSynth- doesn't take a note name, because noise doesn't have any particular pitch. You have to specify only the duration and the
time when the note should be played.

When you play this example, you should hear a stream of hi-hat sounds.

It doesn't sound great yet, because we haven't added the filter.