created a -Tone.Filter- wrapped in _hiHatFilter name varaible
First, we use -Tone.Filter_ to create a  - bandpass filter -. This kind of filer lets through ('passes')
only the frequencies at or around a frequency of your choice. In this case, we tell the filer to pass the
band of frequencies around 15,000 Hz, or 15 kHz, while eliminating all others.

The human range of hearing is roughly 20 Hz to 20 kHz, so our filter lets through only parts of the noise
that are very high in pitch.

Note -- We're connecting the filter to the output, and then connecting the synth to the filter. -- This means
that the synth's sound is run through the filter before being ouput through speakers or headphones.

As a result, when you play this example you should hear the same hi-hat sounds, but with limited to high
frequencies only which sounds a bit more like a real hi-hat.