created a class -Snare-
In the constructor, we start by creating a filter and noise synth. This is similar to how we created the hi-hat,
except we use a frequency of 5,000 Hz for the bandpass filter, to reflect the lower sound of a snare drum.

Next, we create the pitched synth, which uses a similar amplitude envelope to the noise synth but with an even shorter
attack to simulate the sound of a snare drum ( in a real snare drum, the snares are triggered by the vibration of the
drum skin, so they lag behind the sound of the drum slightly ). -- The synth is configured with a sine wave oscillator.

Since we'll be playing the two synths simultaneously, we give each one a volume of -12, which results in an overall volume
similar to the hi-hat.

The -triggerAttackRelease- method takes just a duration and a _when_ parameter. These are passed to the underlying synths'
-triggerAttackRelease- methods. -- When we trigger the pitched synth, we give it a note name of G3, which is the pitch
I decided to tune the snare to. The inclusion of the pitched synth is subtle but makes the drum sound a bit more realistic.

Next, we instantiate the class wrapped in -snare- named variable.

Finally, we create a new -Tone.Loop- object. -- This loop is four times as long as the hi-hat loop ('2n' instead of '8n', or
a half note instead of an eighth note) and starts after one quarter note. -- This means there will be a snare hit on the second
and fourth beats of every bar.

When you play this example, you should hear the hi-hat every eighth note and the snare every two quarter notes.