updated README.md and creatred a -Tone.MembraneSynth-
The options for the -MembraneSynth- include _pitchDecay_ -- which specifies in seconds how quickly the frequency should
change, and _octaves_ -- which specifies how many octaves to drop the frequency in that time.

In our loop we trigger the synth with a frequency of 50 Hz. -- This loop has the same '2n' duration as the snare loop, but
starting at time zero, which means that the kick and snare sounds will alternate every quarter note, giving a classic rock
drum beat.

When you play this example, you might recognize it as the basic drum pattern of a lot of songs.