created a -Tone.Reverb- wrapped in reverb named variable
First we create our Reverb effect. The _decay_ setting describes how long (in seconds) the reverberation will continue after the sound stops.
The higher number, the more echoey the effect. -- The _wet_ setting specifies how much of the reverb sound is passed through compared with the
original sound. - In this case, 0.3 means that the output of this effect will be 30 percent reverb and 70 percent the original sound. The higher
the _wet_ setting, the more prominent the reverb effect will be.

The rest of the changes replace any instances of -toDestination()- with -connect(reverb)-  -- This way all the drum sounds are passed through
the reverb effect before the reverb effect is sent to the output.

When you play this example, the drums should sound more like they're being played in a room