created -mkSequence- function
The -mkSequence- function takes a string like 'x...x...' and converts it to an array of strings and nulls, like
['x', null, null, null, 'x', null, null, null], which is the format we need for -Tone.Sequence. -- It splits the
string into an array of individual characters using the -split- method and uses the arrya -map- method to create
a new array by calling a function for each character.

If the character is '.' -- then it replaces it with a null in the new array. Otherwise, it passes the character
through unchanged.